requirements.txt
newsapi-python
slack-sdk
pytz
python-dateutil